Harmonious ratios

A simple named vector of "harmonious" ratios.

Usage

ratios

Format

A named vector

Details

Ratios fall into the following categories:

chromatic ratios derived from western music theory

golden ratio

"triple" and "double octave" ratios -- so set includes simple ratios of 1:1 (unison), 1:2 (octave), 1:3, and 1:4

All ratios (except for 1:1) appear in pairs, e.g. "major third" (1.25) and "inverse major third" (0.8). Ratios greater than one imply a portrait orientation, and ratios less than one imply a landscape orientation. Vector is sorted from square (unison) to the tallest/widest proportions.

See also

Examples

cbind(ratios)
#>                          ratios
#> unison                1.0000000
#> minor 2nd             1.0666667
#> inverse minor 2nd     0.9375000
#> major 2nd             1.1250000
#> inverse major 2nd     0.8888889
#> minor 3rd             1.2000000
#> inverse minor 3rd     0.8333333
#> major 3rd             1.2500000
#> inverse major 3rd     0.8000000
#> fourth                1.3333333
#> inverse fourth        0.7500000
#> fifth                 1.5000000
#> inverse fifth         0.6666667
#> minor 6th             1.6000000
#> inverse minor 6th     0.6250000
#> golden ratio          1.6180340
#> inverse golden ratio  0.6180340
#> major 6th             1.6666667
#> inverse major 6th     0.6000000
#> minor 7th             1.7777778
#> inverse minor 7th     0.5625000
#> major 7th             1.8750000
#> inverse major 7th     0.5333333
#> octave                2.0000000
#> inverse octave        0.5000000
#> triple                3.0000000
#> inverse triple        0.3333333
#> double octave         4.0000000
#> inverse double octave 0.2500000